* files.make: 'xml_utils/attr_ulong.c is added.
* reuse_exec.h:
(task_NewWait) new function prototype is added.

* run.h:
(struct testinfo) field 'max_memory_used' now has unsigned long type.

* xml_utils.h: copyright notice is updated;
(xml_attr_ulong) new function prototype is added.

* testing_report_xml.h:
(struct testing_report_test) field 'max_memory_used' now has
  unsigned long type.

* master_html.c:
(write_xml_testing_report) format spec for 'max_memory_used'
  field is updated.

* reports.c: copyright notice is updated;
(write_xml_tex_testing_report) format spec for 'max_memory_used'
  is updated.

* testing_report_xml.c:
(parse_test) "max-memory-used" attribute is parsed to the
  unsigned long field 'max_memory_used';
(testing_report_unparse_xml) unparsing of 'max_memory_used'
  is updated.

+++ run.h
@@ -35,7 +35,7 @@ struct testinfo
   int            max_score;     /* maximal score for this test */
   long           times;         /* execution time */
   long           real_time;     /* execution real time */
-  int            max_memory_used;
+  unsigned long  max_memory_used;
   char          *input;         /* the input */
   long           input_size;
   int            has_input_digest;
